requirements.txt
Groq
streamlit
Pillow
langsmith